requirements.txt
fastapi>=0.68.0,<0.70.0
uvicorn>=0.15.0,<0.17.0
torch>=1.10.0,<2.0.0
torchvision>=0.11.0,<0.13.0
Pillow>=8.4.0,<9.0.0
numpy>=1.21.0,<1.24.0
python-multipart
scikit-learn>=1.2.2
